fix: Issue #101 - Fix Line Drawing Mode slope accumulator overflow detection
The Line Drawing Mode slope accumulator was incorrectly implemented:
- slopeAccum was uint16_t, which can never be > 0xFFFF
- Overflow check used > 0xFFFF with & 0xFFFF masking (nonsensical for 16-bit)
- This prevented proper Bresenham-like slope accumulation

Fix:
- Changed slopeAccum to uint32_t to properly accumulate slope values
- Updated overflow check to >= 0x10000 (proper 16-bit carry detection)
- Changed masking to -= 0x10000 (proper carry subtraction)

This fixes the slope accumulator logic that determines when to step the
minor axis in line drawing mode. The accumulator now properly detects when
it crosses a 16-bit boundary (0x10000) and triggers the minor axis step.

Impact:
- Line Drawing Mode slope accumulation now works correctly
- Should fix accuracy issues in dmagic_line_unit_test
- Better hardware compliance with real MEGA65 line drawing

+++ src/plugins/devices/f018b_dma/main/f018b_dma.h
@@ -159,7 +159,7 @@ private:
         uint32_t xCol;         // X card boundary offset (options $87/$88 or $97/$98)
         uint32_t yCol;         // Y card boundary offset (options $89/$8A or $99/$9A)
         uint16_t slope;        // Slope value added to accumulator each step ($8B/$8C or $9B/$9C)
-        uint16_t slopeAccum;   // Slope accumulator ($8D/$8E or $9D/$9E)
+        uint32_t slopeAccum;   // Slope accumulator (32-bit to properly detect overflow at 0x10000)
         uint8_t  slopeType;    // Control byte ($8F or $9F): bit7=enable, bit6=Y-major, bit5=negative
     };
     LineMode m_srcLine, m_dstLine;
